added hoc for map component
